updated makeContact.js
